Update HNIDS.ps1
Added misc features.. testing idea of adding a scheduled task that runs daily, tracks progress, and outputs a CSV file when it's done.  Will talk with people about what is better.. using Out-Gridview or just quietly exporting to a CSV and running daily at 8PM or whatever time is best.